Time: 73 ms (97.79%), Space: 116.7 MB (47.55%) - LeetHub

+++ 2028-find-missing-observations/2028-find-missing-observations.cpp
@@ -4,28 +4,25 @@ public:
         std::ios_base::sync_with_stdio(false);
         std::cin.tie(nullptr);
         std::cout.tie(nullptr);
-        
-        long sum = 0, x = 0, m = rolls.size(), val = 0, rem = 0;
+
+        int sum = 0, x = 0, m = rolls.size(), val = 0, rem = 0;
 
         sum = accumulate(rolls.begin(), rolls.end(), 0);
-        
         x = mean * (n + m) - sum;
         val = x / n;
 
-        if(val > 6 || val < 1)
+        if (val > 6 || val < 1)
             return {};
 
         vector<int> res(n, val);
-
         rem = x - val * n;
 
-        if(val == 6 && rem != 0)
+        if (val == 6 && rem != 0)
             return {};
 
-        for(int i = 0; i < rem; i++)
+        for (int i = 0; i < rem; i++)
             res[i]++;
 
         return res;
-
     }
 };
